Add option to configure expire threshold
The '-e' option takes and argument in seconds and ignores all keys
that expire within this threshold.

+++ tasks/process_metabuf_task.cc
@@ -61,7 +61,8 @@ void ProcessMetabufTask::ProcessMetaBuffer(MetaBufferSlice* mslice) {
       abort();
     }
 
-    if (expiry != -1 && MemcachedUtils::KeyExpiresSoon(now, expiry)) {
+    if (expiry != -1 && MemcachedUtils::KeyExpiresSoon(now,
+        static_cast<uint32_t>(expiry))) {
       owning_thread()->increment_keys_ignored();
       continue;
     }
